Updating C Extension

+++ ml/optimizer/gradient_descent.c
@@ -1,6 +1,6 @@
 #include <python3.6/Python.h>
 
-int call_func(PyObject* func, int* theta, int num_theta);
+int call_func(PyObject* func, double* theta, int num_theta);
 
 int * _optimize(PyObject* func, double learning_rate, int steps, double* init_theta,  
                 double dx, int num_theta)
@@ -69,7 +69,9 @@ static PyObject * optimize(PyObject* self, PyObject* args)
     printf("In the function!\n");
     if (!PyArg_ParseTuple(args, "OiiOii", &func, &learning_rate, &steps, &init_theta, &dx, &num_theta))
         return NULL;
+    
     int ret = call_func(func, theta, num_theta);
+
     //double test = PyList_GetItem(init_theta, 0);
     
     return Py_BuildValue("i", ret);
